// src/app/Models/Fillingweekandmonth.ts
export class WeekandMonth_ {
    Transaction_ID: number | undefined;
    Sort_Date: Date | undefined;
    Amount_Filling_Time_diesel_: number | undefined;
    Amount_Filling_Time_gasohol95_: number | undefined;
    Amount_of_Fuel_diesel_: number | undefined;
    Amount_of_Fuel_gasohol95_: number | undefined;
    C_Truck_In: number | undefined;
    Truck_Out: number | undefined;

}
// export interface UsersAll{
//      result : Customer_[]
//      message : string

// }